+++ service/controller/resource/prometheus/resource.go
@@ -114,12 +114,6 @@ func toPrometheus(v interface{}, createPVC bool, storageSize resource.Quantity) 
 			},
 			Replicas: &replicas,
 			Resources: corev1.ResourceRequirements{
-				Limits: corev1.ResourceList{
-					// cpu: 100m
-					corev1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI),
-					// memory: 100Mi
-					corev1.ResourceMemory: *resource.NewQuantity(1*1024*1024*1024, resource.BinarySI),
-				},
 				Requests: corev1.ResourceList{
 					// cpu: 100m
 					corev1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI),