+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Moved prometheus storage from `emptyDir` to a `persistentVolumeClaim`
+- Remove tenant cluster prometheus limits
 
 ## [0.3.2] - 2020-07-24